Regular upmerge.

+++ scripts/pylib/twister/twisterlib/environment.py
@@ -72,7 +72,7 @@ Artificially long but functional example:
 
     run_group_option = parser.add_mutually_exclusive_group()
 
-    device = parser.add_mutually_exclusive_group(required="--device-testing" in sys.argv)
+    device = parser.add_mutually_exclusive_group()
 
     test_or_build = parser.add_mutually_exclusive_group()
 
@@ -603,9 +603,21 @@ structure in the main Zephyr tree: boards/<arch>/<board_name>/""")
         """)
 
     parser.add_argument(
+        "--report-summary", action="store", nargs='?', type=int, const=0,
+        help="Show failed/error report from latest run. Default shows all items found. "
+             "However, you can specify the number of items (e.g. --report-summary 15). "
+             "It also works well with the --outdir switch.")
+
+    parser.add_argument(
         "--report-suffix",
         help="""Add a suffix to all generated file names, for example to add a
         version or a commit ID.
+        """)
+
+    parser.add_argument(
+        "--report-all-options", action="store_true",
+        help="""Show all command line options applied, including defaults, as
+        environment.options object in twister.json. Default: show only non-default settings.
         """)
 
     parser.add_argument(
@@ -738,7 +750,7 @@ structure in the main Zephyr tree: boards/<arch>/<board_name>/""")
     return parser
 
 
-def parse_arguments(parser, args, options = None):
+def parse_arguments(parser, args, options = None, on_init=True):
     if options is None:
         options = parser.parse_args(args)
 
@@ -782,7 +794,7 @@ def parse_arguments(parser, args, options = None):
     if options.coverage:
         options.enable_coverage = True
 
-    if not options.coverage_platform:
+    if options.enable_coverage and not options.coverage_platform:
         options.coverage_platform = options.platform
 
     if options.coverage_formats:
@@ -794,6 +806,10 @@ def parse_arguments(parser, args, options = None):
 
     if options.enable_valgrind and not shutil.which("valgrind"):
         logger.error("valgrind enabled but valgrind executable not found")
+        sys.exit(1)
+
+    if (not options.device_testing) and (options.device_serial or options.device_serial_pty or options.hardware_map):
+        logger.error("Use --device-testing with --device-serial, or --device-serial-pty, or --hardware-map.")
         sys.exit(1)
 
     if options.device_testing and (options.device_serial or options.device_serial_pty) and len(options.platform) != 1:
@@ -856,13 +872,13 @@ def parse_arguments(parser, args, options = None):
         # Strip off the initial "--" following validation.
         options.extra_test_args = options.extra_test_args[1:]
 
-    if not options.allow_installed_plugin and PYTEST_PLUGIN_INSTALLED:
+    if on_init and not options.allow_installed_plugin and PYTEST_PLUGIN_INSTALLED:
         logger.error("By default Twister should work without pytest-twister-harness "
                      "plugin being installed, so please, uninstall it by "
                      "`pip uninstall pytest-twister-harness` and `git clean "
                      "-dxf scripts/pylib/pytest-twister-harness`.")
         sys.exit(1)
-    elif options.allow_installed_plugin and PYTEST_PLUGIN_INSTALLED:
+    elif on_init and options.allow_installed_plugin and PYTEST_PLUGIN_INSTALLED:
         logger.warning("You work with installed version of "
                        "pytest-twister-harness plugin.")
 
@@ -874,12 +890,13 @@ def strip_ansi_sequences(s: str) -> str:
 
 class TwisterEnv:
 
-    def __init__(self, options=None) -> None:
+    def __init__(self, options=None, default_options=None) -> None:
         self.version = "Unknown"
         self.toolchain = None
         self.commit_date = "Unknown"
         self.run_date = None
         self.options = options
+        self.default_options = default_options
 
         if options and options.ninja:
             self.generator_cmd = "ninja"
@@ -913,6 +930,18 @@ class TwisterEnv:
         self.test_config = options.test_config if options else None
 
         self.alt_config_root = options.alt_config_root if options else None
+
+    def non_default_options(self) -> dict:
+        """Returns current command line options which are set to non-default values."""
+        diff = {}
+        if not self.options or not self.default_options:
+            return diff
+        dict_options = vars(self.options)
+        dict_default = vars(self.default_options)
+        for k in dict_options.keys():
+            if k not in dict_default or dict_options[k] != dict_default[k]:
+                diff[k] = dict_options[k]
+        return diff
 
     def discover(self):
         self.check_zephyr_version()
